+++ js/src/test/scala/com/thirdparty/defs/attrs/SvgAttrs.scala
@@ -983,8 +983,8 @@ trait SvgAttrs {
   lazy val xmlSpace: SvgAttr[String] = stringSvgAttr("space", namespace = "xml")
 
 
-  /** @see https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/xmlns */
-  lazy val xmlns: SvgAttr[String] = stringSvgAttr("xmlns", namespace = "xmlns")
+  /** @see https://developer.mozilla.org/en-US/docs/Web/SVG/Element/svg */
+  lazy val xmlns: SvgAttr[String] = stringSvgAttr("xmlns")
 
 
   /** @see https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/xmlns:xlink */